Changed id attribute type for consistency

+++ src/preprocessing/passes/normalize.h
@@ -37,14 +37,14 @@ struct NodeInfo
   /** Compressed string encoding of the DAG, aka patterns. */
   std::string encoding;
   /** Equivalence class ID. */
-  uint32_t equivClass;
+  size_t equivClass;
   /** First occurrence index (counter) of each symbol when traversing the DAG,
    * aka super-pattern */
   std::unordered_map<std::string, int32_t> role;
   /** Variable names and their indices. */
   std::vector<std::pair<std::string, int32_t>> varNames;
   /** Index of the assertion after first round of sorting. */
-  uint32_t id;
+  size_t id;
 
   /** Default constructor. */
   NodeInfo() {}
